numeric -> integer

+++ sql/4.5/orcadbup_syunou_main_4.5.0-1.sql
@@ -1,7 +1,7 @@
 \set ON_ERROR_STOP
 
-alter table tbl_syunou_main add column chosei1 numeric(7,0);
-alter table tbl_syunou_main add column chosei2 numeric(7,0);
+alter table tbl_syunou_main add column chosei1 integer;
+alter table tbl_syunou_main add column chosei2 integer;
 update tbl_syunou_main set chosei1 = chosei;
 update tbl_syunou_main set chosei2 = 0 ;
