+++ sql/4.5/orcadbup_syutotal_4.5.0-1.sql
@@ -1,7 +1,7 @@
 \set ON_ERROR_STOP
 
-alter table tbl_syutotal add column chosei1 numeric(7,0);
-alter table tbl_syutotal add column chosei2 numeric(7,0);
+alter table tbl_syutotal add column chosei1 integer;
+alter table tbl_syutotal add column chosei2 integer;
 update tbl_syutotal set chosei1 = chosei;
 update tbl_syutotal set chosei2 = 0 ;
